Setup performance test on github actions (#853)
Setup performance test on github actions (Closes #846)

* Setup github action to import and run model from Fröhlich et al., Cell Systems, 2018 (forward simulation, forward+adjoint sensitivities) 
* Don't require use of ModelPtr.get to call ExpData(Model)
* Fix import in generated model Python package
* Setup AMICI standalone scripts as setuptools entrypoints
* Add reference values for wall times and check we are within the limits
