+++ tests/performance/reference.yml
@@ -1,0 +1,8 @@
+# Reference wall times (seconds) with some buffer
+create_sdist: 25
+install_sdist: 120
+petab_import: 1560 # 26'
+install_model: 400
+forward_simulation: 2
+forward_sensitivities: 6
+adjoint_sensitivities: 120 # TODO reduce when #849 fixed
